+++ scheduler.js
@@ -1,0 +1,3 @@
+import {weekly_timetable, subjects} from './tt'
+
+console.log(weekly_timetable)
