+++ tt.js
@@ -1,0 +1,177 @@
+// working on timetabling Offsite
+
+console.log('trying timetable in javascript')
+
+let weekly_timetable = {
+    mon:{
+       '9X': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+       '9Y': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+       '9Z': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' }
+    },
+    tue:{
+        '9X': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+        '9Y': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+        '9Z': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' }
+     },
+    wed:{
+        '9X': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+        '9Y': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+        '9Z': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' }
+     },
+     thur:{
+        '9X': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+        '9Y': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+        '9Z': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' }
+     },
+    fri:{
+        '9X': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+        '9Y': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' },
+        '9Z': { 1:'free',2:'free',3:'free',4:'free',5:'free',6:'free',7:'free',8:'free' }
+     }
+    
+}
+
+const subjects = [
+    {eng: {
+        daily:[0,7],
+        weekly:[0,12],
+        teacher: 'BA',
+        code:'ENG(BA)',
+        mon:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },
+        tue:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },
+        wed:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },        
+        thur:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },        
+        fri:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        }
+    },
+    math: {
+        daily:[0,7],
+        weekly:[0,12],
+        teacher: 'AT',
+        code:'MATH(AT)',
+        mon:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },
+        tue:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },
+        wed:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },        
+        thur:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },        
+        fri:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        }
+    },
+    phy: {
+        daily:[0,7],
+        weekly:[0,12],
+        teacher: 'OJ',
+        code:'PHY(OJ)',
+        mon:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },
+        tue:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },
+        wed:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },        
+        thur:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },        
+        fri:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        }
+    },
+    che: {
+        daily:[0,7],
+        weekly:[0,12],
+        teacher: 'RC',
+        code:'CHE(RC)',
+        wed:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },        
+        thur:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },        
+        fri:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        }
+    },
+    mus: {
+        daily:[0,7],
+        weekly:[0,12],
+        teacher: 'DA',
+        code:'MUS(DA)',
+        mon:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },
+        tue:{
+            '9x':[1,2,3,4,5,6,7,8],
+            '9y':[1,2,3,4,5,6,7,8],
+            '9z':[1,2,3,4,5,6,7,8],
+        },
+        wed:{
+            '9x':[1,2,3,4],
+            '9y':[1,2,3,4],
+            '9z':[1,2,3,4],
+        },        
+        }
+    }
+]
+
+
+
+// console.log(Object.keys(weekly_timetable.mon))
+
+// console.log(Object.keys(weekly_timetable))
